Reorganize plugin code into separate files
