+++ resources/lib/fcast_plugin/main.py
@@ -9,7 +9,10 @@ import selectors
 from urllib.parse import urlparse
 from pathlib import Path
 from base64 import b64encode
-from FCastSession import Event, FCastSession, PlayMessage, PlayBackUpdateMessage, PlayBackState, SeekMessage, SetVolumeMessage, VolumeUpdateMessage
+
+from .FCastSession import Event, FCastSession, PlayMessage, PlayBackUpdateMessage, PlayBackState, SeekMessage, SetVolumeMessage, VolumeUpdateMessage
+from .player import FCastPlayer
+from .util import log_and_notify, debounce
 
 session_threads: List[Thread] = []
 sessions: List[FCastSession] = []
@@ -19,90 +22,9 @@ FCAST_PORT = 46899
 FCAST_TIMEOUT = 60 * 1000
 FCAST_BUFFER_SIZE = 32000
 
-# Retrieve Kodi addon information
-addon       = xbmcaddon.Addon()
-addonname   = addon.getAddonInfo('name')
-
 plugin_handle = int(sys.argv[1]) if len(sys.argv) > 1 else None
 
 player_thread: Thread = None
-
-# Trottle repeated attempts at a function call
-def debounce(func, wait):
-    def debounced(*args, **kwargs):
-        debounced.timer.cancel()
-        debounced.timer = Timer(wait, func, args=args, kwargs=kwargs)
-        debounced.timer.start()
-
-    debounced.timer = Timer(0, lambda: None)  # Initial dummy timer
-    return debounced
-
-class FCastPlayer(xbmc.Player):
-    playback_speed: float = 1.0
-    sessions: List[FCastSession]
-    is_paused: bool = False
-    # Used to perform time updates
-    prev_time: int = -1
-
-    def __init__(self, sessions: FCastSession):
-        self.sessions = sessions
-        super().__init__(self)
-    
-    def doPause(self) -> None:
-        if not self.is_paused:
-            self.is_paused = True
-            self.pause()
-    
-    def doResume(self) -> None:
-        if self.is_paused:
-            self.is_paused = False
-            self.pause()
-
-    def onAVStarted(self) -> None:
-        log_and_notify(addonname, "Playback started")
-        self.is_paused = False
-        # Start time loop once the player is active
-        self.onPlayBackTimeChanged()
-
-    def onPlayBackStopped(self) -> None:
-        self.onPlayBackEnded()
-
-    def onPlayBackPaused(self) -> None:
-        self.is_paused = True
-        self.onPlayBackTimeChanged()
-
-    def onPlayBackResumed(self) -> None:
-        self.is_paused = False
-    
-    def onPlayBackEnded(self) -> None:
-        for session in self.sessions:
-            session.send_playback_update(PlayBackUpdateMessage(
-                0,
-                PlayBackState.IDLE,
-            ))
-    
-    def onPlayBackError(self) -> None:
-        self.onPlayBackEnded()
-    
-    def onPlayBackSpeedChanged(self, speed: int) -> None:
-        self.playback_speed = speed
-    
-    # Not overriden
-    def onPlayBackTimeChanged(self) -> None:
-        time_int = int(self.getTime())
-        self.prev_time = int(self.getTime())
-        pb_message = PlayBackUpdateMessage(
-            time_int,
-            PlayBackState.PAUSED if self.is_paused else PlayBackState.PLAYING,
-        )
-        for session in self.sessions:
-            session.send_playback_update(pb_message)
-    
-    def addSession(self, session: FCastSession):
-        self.sessions.append(session)
-    
-    def removeSession(self, session: FCastSession):
-        self.sessions.remove(session)
     
 # Player needs to be a global so it stays in scope and doesn't get GC'd
 player: FCastPlayer = None
@@ -112,7 +34,7 @@ seeks: list[float] = []
 
 def check_player():
     global player
-    log_and_notify(addonname, "Starting player thread", notify=False)
+    log_and_notify("Starting player thread", notify=False)
     monitor = xbmc.Monitor()
     while not monitor.abortRequested():
         if player.isPlaying():
@@ -122,16 +44,10 @@ def check_player():
         
         if monitor.waitForAbort(0.05):
             break
-    log_and_notify(addonname, "Exiting player thread", notify=False)
-
-# Helper function to both print a message to the Kodi logs and create a notification
-def log_and_notify(tag, msg, icon=xbmcgui.NOTIFICATION_INFO, timeout=3000, loglevel=xbmc.LOGDEBUG, notify=True):
-    xbmc.log("%s: %s" % (tag, msg), level=loglevel)
-    if notify:
-        xbmcgui.Dialog().notification(tag, msg, icon, timeout, True)
+    log_and_notify("Exiting player thread", notify=False)
 
 def handle_play(session: FCastSession, message: PlayMessage):
-    log_and_notify(addonname, f"Client request play", notify=False)
+    log_and_notify(f"Client request play", notify=False)
     play_item: xbmcgui.ListItem = None
     url: str = ''
     play_item = xbmcgui.ListItem()
@@ -140,7 +56,7 @@ def handle_play(session: FCastSession, message: PlayMessage):
         parsed_url = urlparse(url)
         # Detect HLS stream
         if Path(parsed_url.path).suffix == '.m3u8':
-            log_and_notify(addonname, 'Detected HLS stream', notify=False)
+            log_and_notify('Detected HLS stream', notify=False)
             # Use inputstream adaptive to handle HLS stream
             play_item.setContentLookup(False)
             play_item.setMimeType('application/x-mpegURL')
@@ -176,7 +92,7 @@ def do_seek():
 
 def handle_seek(session: FCastSession, message: SeekMessage):
     global player, seeks
-    log_and_notify(addonname, f"Client request seek to {message.time}", notify=False)
+    log_and_notify(f"Client request seek to {message.time}", notify=False)
     # Send FCastMessage so the client's seek bar position updates better
     session.send_playback_update(PlayBackUpdateMessage(
         message.time,
@@ -190,22 +106,22 @@ def handle_seek(session: FCastSession, message: SeekMessage):
 
 def handle_stop(session: FCastPlayer, message = None):
     global player
-    log_and_notify(addonname, f"Client request stop", notify=False)
+    log_and_notify(f"Client request stop", notify=False)
     player.stop()
 
 def handle_pause(session: FCastPlayer, message = None):
     global player
-    log_and_notify(addonname, f"Client request pause", notify=False)
+    log_and_notify(f"Client request pause", notify=False)
     player.doPause()
 
 def handle_resume(session: FCastPlayer, message = None):
     global player
-    log_and_notify(addonname, f"Client request resume", notify=False)
+    log_and_notify(f"Client request resume", notify=False)
     player.doResume()
 
 def handle_volume(session: FCastSession, message: SetVolumeMessage):
     global player
-    log_and_notify(addonname, f"Client request set volume at {message.volume}", notify=False)
+    log_and_notify(f"Client request set volume at {message.volume}", notify=False)
     volume_level = int(message.volume * 100)
     xbmc.executebuiltin(f'SetVolume({volume_level})')
 
@@ -214,7 +130,7 @@ def connection_handler(conn: socket.socket, addr):
     global player
 
     monitor = xbmc.Monitor()
-    log_and_notify(addonname, "Connection from %s" % addr[0])
+    log_and_notify("Connection from %s" % addr[0])
 
     session = FCastSession(conn)
 
@@ -247,12 +163,12 @@ def connection_handler(conn: socket.socket, addr):
 
     player.removeSession(session)
     session.close()
-    log_and_notify(addonname, "Connection closed from %s" % addr[0])
+    log_and_notify("Connection closed from %s" % addr[0])
 
 def main():
     global player, sessions, session_threads, player_thread
 
-    log_and_notify(addonname, "Starting FCast receiver ...")
+    log_and_notify("Starting FCast receiver ...")
     # List of active sessions
 
     # Create a socket for the FCast receiver
@@ -269,7 +185,7 @@ def main():
         s.bind((FCAST_HOST, FCAST_PORT))
         s.listen()
     except:
-        log_and_notify(addonname, "Bind failed", xbmcgui.NOTIFICATION_ERROR)
+        log_and_notify("Bind failed", xbmcgui.NOTIFICATION_ERROR)
         s.close()
         exit()
 
@@ -277,7 +193,7 @@ def main():
     selector = selectors.DefaultSelector()
     selector.register(s, selectors.EVENT_READ, data=None)
 
-    log_and_notify(addonname, "Server listening on port %d" % FCAST_PORT, timeout=1000)
+    log_and_notify("Server listening on port %d" % FCAST_PORT, timeout=1000)
 
     monitor = xbmc.Monitor()
     # Loop for new connections
@@ -302,7 +218,7 @@ def main():
 
     s.close()
 
-    log_and_notify(addonname, "Server stopped")
+    log_and_notify("Server stopped")
     exit()
 
 if __name__ == '__main__':
