+++ resources/lib/startup.py
@@ -1,0 +1,4 @@
+from fcast_plugin.main import main
+
+# Run plugin
+main()
